requirements.txt
Django==4.2.6
djangorestframework==3.14.0
geopy==2.4.0
numpy==1.26.0
pandas==2.1.1
gunicorn==21.2.*
django-multiselectfield==0.1.12
